fix:
Error on line 109 in C:\projects\fife-windows-sdk\installer\fife-sdk.iss: Parameter "Name" includes invalid characters.
It may only include alphanumeric characters, underscores, slashes (/), and/or backslashes (\), may not start with a number and may not start or end with a slash or a backslash. Names 'not', 'and' and 'or' are reserved.
Compile aborted.